+++ packages/ricos-content/web/lib/modifier-infra.ts
@@ -1,0 +1,2 @@
+export { getModifier, Modifier } from '../src/modifier-infra';
+export { getExtractor, Extractor } from '../src/extractor-infra';