+++ packages/plugin-gallery/web/src/editor.ts
@@ -4,13 +4,17 @@ import { ModalsMap } from './modals';
 import { DEFAULTS } from './defaults';
 import { EditorPluginCreator } from 'wix-rich-content-common';
 import { createGalleryData } from './createGalleryData';
+import { createRicosExtensions } from './tiptap';
+import { TiptapEditorPlugin } from 'wix-tiptap-editor';
 
 export const pluginGallery: EditorPluginCreator<GalleryPluginEditorConfig> = config => {
+  const pluginConfig = { ...DEFAULTS.config, ...config };
   return {
-    config: { ...DEFAULTS.config, ...config },
+    config: pluginConfig,
     type: GALLERY_TYPE,
     createPlugin: createGalleryPlugin,
     ModalsMap,
     createPluginData: createGalleryData,
-  };
+    tiptapExtensions: createRicosExtensions(pluginConfig),
+  } as TiptapEditorPlugin;
 };