+++ packages/ricos-content/web/src/velo-adapter/extract.ts
@@ -1,34 +1,50 @@
-import { RichContent, Node_Type, TextData, Node, Media } from 'ricos-schema';
+import * as A from 'fp-ts/Array';
+import { pipe } from 'fp-ts/function';
+import * as M from 'fp-ts/Monoid';
+import { fromNullable } from 'fp-ts/Option';
+import {
+  GalleryData,
+  ImageData,
+  Media,
+  Node_Type,
+  RichContent,
+  TextData,
+  VideoData,
+} from 'ricos-schema';
 import { RICOS_NODE_TYPE_TO_DATA_FIELD } from '../consts';
 import { extract } from '../RicosContentAPI/extract';
 
-type PluginData = NonNullable<
-  Node[typeof RICOS_NODE_TYPE_TO_DATA_FIELD[keyof typeof RICOS_NODE_TYPE_TO_DATA_FIELD]]
->;
-
-export const getText = (content: RichContent) =>
+export const getText = (content: RichContent): TextData[] =>
   extract(content.nodes)
-    .filter(node => node.type === Node_Type.TEXT)
-    .map(node => node.textData)
-    .filter(data => !!data?.text)
-    .get() as TextData[];
+    .map(({ textData }) => (textData?.text ? textData : ((null as unknown) as TextData)))
+    .get();
 
-export const getPluginData = (content: RichContent, type: Node_Type) =>
+export const getPluginData = <T>(content: RichContent, type: Node_Type) =>
   extract(content.nodes)
-    .filter(node => node.type === type)
-    .map(node => node[RICOS_NODE_TYPE_TO_DATA_FIELD[node.type]])
-    .get() as PluginData[];
+    .map(node => node[RICOS_NODE_TYPE_TO_DATA_FIELD[type]])
+    .get() as T[];
 
-const extractNodeMedia = (node: Node) =>
-  node.type === Node_Type.IMAGE
-    ? [node.imageData?.image]
-    : node.type === Node_Type.VIDEO
-    ? [node.videoData?.video]
-    : [node.galleryData?.items.map(item => item.image?.media || item.video?.media)].flat();
+const MediaM = A.getMonoid<Media>();
 
-export const getMedia = (content: RichContent) =>
-  extract(content.nodes)
-    .filter(node => [Node_Type.IMAGE, Node_Type.VIDEO, Node_Type.GALLERY].includes(node.type))
-    .get()
-    .flatMap(extractNodeMedia)
-    .filter(media => !!media) as Media[];
+export const getMedia = (content: RichContent): Media[] =>
+  M.concatAll(MediaM)([
+    pipe(
+      getPluginData<ImageData>(content, Node_Type.IMAGE),
+      A.map(({ image }) => image),
+      A.map(fromNullable),
+      A.compact
+    ),
+    pipe(
+      getPluginData<VideoData>(content, Node_Type.VIDEO),
+      A.map(({ video }) => video),
+      A.map(fromNullable),
+      A.compact
+    ),
+    pipe(
+      getPluginData<GalleryData>(content, Node_Type.GALLERY),
+      A.chain(({ items }) => items),
+      A.map(({ image, video }) => image?.media || video?.media),
+      A.map(fromNullable),
+      A.compact
+    ),
+  ]);