+++ packages/tiptap-editor/web/src/components/RichContentAdapter/RichContentAdapter.tsx
@@ -9,18 +9,57 @@ import {
   TranslationFunction,
   EditorPlugin,
   TextAlignment,
+  GALLERY_TYPE,
+  RICOS_GALLERY_TYPE,
+  FILE_UPLOAD_TYPE,
+  RICOS_FILE_TYPE,
+  GIPHY_TYPE,
+  RICOS_GIPHY_TYPE,
+  VIDEO_TYPE,
+  RICOS_VIDEO_TYPE,
+  RICOS_TEXT_COLOR_TYPE,
+  RICOS_TEXT_HIGHLIGHT_TYPE,
 } from 'wix-rich-content-common';
-import { toTiptap } from '../../converters';
+import { toTiptap } from '../../content-utils';
 import { Editor } from '@tiptap/core';
-
-const TIPTAP_DIVIDER_TYPE = 'divider'; //should be taken from tip-tap common
-const TIPTAP_IMAGE_TYPE = 'image';
+import {
+  HEADER_BLOCK,
+  UNSTYLED,
+  BULLET_LIST_TYPE,
+  CODE_BLOCK_TYPE,
+  BLOCKQUOTE,
+  HEADINGS_TYPE,
+  NUMBERED_LIST_TYPE,
+  TIPTAP_DIVIDER_TYPE,
+  TIPTAP_IMAGE_TYPE,
+  TIPTAP_GALLERY_TYPE,
+  TIPTAP_FILE_TYPE,
+  TIPTAP_GIF_TYPE,
+  TIPTAP_VIDEO_TYPE,
+} from 'ricos-content';
 
 const PLUGIN_TYPE_MAP = {
   [RICOS_DIVIDER_TYPE]: TIPTAP_DIVIDER_TYPE,
   [DIVIDER_TYPE]: TIPTAP_DIVIDER_TYPE,
   [RICOS_IMAGE_TYPE]: TIPTAP_IMAGE_TYPE,
   [IMAGE_TYPE]: TIPTAP_IMAGE_TYPE,
+  [GALLERY_TYPE]: TIPTAP_GALLERY_TYPE,
+  [RICOS_GALLERY_TYPE]: TIPTAP_GALLERY_TYPE,
+  [FILE_UPLOAD_TYPE]: TIPTAP_FILE_TYPE,
+  [RICOS_FILE_TYPE]: TIPTAP_FILE_TYPE,
+  [GIPHY_TYPE]: TIPTAP_GIF_TYPE,
+  [RICOS_GIPHY_TYPE]: TIPTAP_GIF_TYPE,
+  [VIDEO_TYPE]: TIPTAP_VIDEO_TYPE,
+  [RICOS_VIDEO_TYPE]: TIPTAP_VIDEO_TYPE,
+};
+
+const headingTypeToLevelMap = {
+  'header-one': 1,
+  'header-two': 2,
+  'header-three': 3,
+  'header-four': 4,
+  'header-five': 5,
+  'header-six': 6,
 };
 
 // todo : should change to RichContentInterface
@@ -92,10 +131,15 @@ export class RichContentAdapter implements TiptapAPI {
       }),
 
       insertDecoration: (type, data) => {
-        if (type !== RICOS_LINK_TYPE) {
-          console.error(`${type} decoration not supported`);
+        const decorationCommandMap = {
+          [RICOS_LINK_TYPE]: data => this.editor.commands.setLink({ link: data }),
+          [RICOS_TEXT_COLOR_TYPE]: data => this.editor.commands.setColor(data.color),
+          [RICOS_TEXT_HIGHLIGHT_TYPE]: data => this.editor.commands.setHighlight(data.color),
+        };
+        if (decorationCommandMap[type]) {
+          decorationCommandMap[type](data);
         } else {
-          this.editor.commands.setLink({ link: data });
+          console.error(`${type} decoration not supported`);
         }
       },
       hasLinkInSelection: () => {
@@ -135,6 +179,33 @@ export class RichContentAdapter implements TiptapAPI {
           });
         });
         return link;
+      },
+      setBlockType: type => {
+        const blockTypeCommandMap = {
+          [UNSTYLED]: () => this.editor.commands.setParagraph(),
+          [HEADINGS_TYPE]: level => this.editor.commands.toggleHeading({ level }),
+          [BLOCKQUOTE]: () => this.editor.commands.toggleBlockquote(),
+          [CODE_BLOCK_TYPE]: () => this.editor.commands.toggleCodeBlock(),
+          [BULLET_LIST_TYPE]: () => this.editor.commands.toggleBulletList(),
+          [NUMBERED_LIST_TYPE]: () => this.editor.commands.toggleOrderedList(),
+        };
+        if (Object.values(HEADER_BLOCK).includes(type)) {
+          blockTypeCommandMap.headings(headingTypeToLevelMap[type]);
+        } else if (blockTypeCommandMap[type]) {
+          blockTypeCommandMap[type]();
+        } else {
+          console.error(`${type} block type not supported`);
+        }
+      },
+      deleteDecoration: type => {
+        const deleteDecorationCommandMap = {
+          [RICOS_LINK_TYPE]: () => this.editor.commands.unsetLink(),
+        };
+        if (deleteDecorationCommandMap[type]) {
+          deleteDecorationCommandMap[type]();
+        } else {
+          console.error(`delete ${type} decoration type not supported`);
+        }
       },
     };
   }
@@ -179,7 +250,6 @@ export class RichContentAdapter implements TiptapAPI {
     saveSelectionState: () => {},
     loadSelectionState: () => {},
     triggerDecoration: () => {},
-    deleteDecoration: () => {},
     setBlock: () => {},
     deleteBlock: () => {},
     undo: () => {},