+++ packages/tiptap-editor/web/src/index.ts
@@ -2,4 +2,4 @@ export { TiptapAPI, CreateRicosExtensions } from './types';
 export { TiptapEditorPlugin } from './models/extension-types';
 export { RicosTiptapEditor } from './components/RicosTiptapEditor';
 export { RichContentAdapter } from './components/RichContentAdapter';
-export * from './converters';
+export * from './content-utils';